fix pipeline

+++ aui.views/src/AUI/View/AAbstractTextField.cpp
@@ -179,7 +179,7 @@ void AAbstractTextField::invalidateFont() {
 }
 
 void AAbstractTextField::onCharEntered(wchar_t c) {
-    mCursorIndex = std::min(mCursorIndex, mContents.size());
+    mCursorIndex = std::min(mCursorIndex, static_cast<unsigned int> (mContents.size()));
     if (c == '\n' || c == '\r')
         return;
     AView::onCharEntered(c);
